requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
charset-normalizer==3.2.0
frozenlist==1.4.0
html5tagger==1.3.0
httptools==0.6.0
idna==3.4
multidict==6.0.4
numpy==1.25.2
opencv-python==4.8.0.76
Pillow==10.0.0
sanic==23.6.0
sanic-routing==23.6.0
tracerite==1.1.0
typing_extensions==4.7.1
websockets==11.0.3
yarl==1.9.2
